refactor: use BigInteger instead of BigDecimal for amounts
Store all balance/amount values in the smallest unit (e.g., cents),
eliminating the need for decimal precision. Changed from BigDecimal
to BigInteger across all modules:

- shared: DomainEvent, Commands, Queries DTOs
- command: AccountAggregate, AccountCommandController
- query: AccountSummary model, V1 migration (DECIMAL -> BIGINT)

+++ shared/src/main/kotlin/com/example/strata/shared/dto/Commands.kt
@@ -1,22 +1,22 @@
 package com.example.strata.shared.dto
 
-import java.math.BigDecimal
+import java.math.BigInteger
 import java.util.UUID
 
 data class CreateAccountCommand(
     val accountId: UUID = UUID.randomUUID(),
     val accountName: String,
-    val initialBalance: BigDecimal = BigDecimal.ZERO
+    val initialBalance: BigInteger = BigInteger.ZERO
 )
 
 data class DepositMoneyCommand(
     val accountId: UUID,
-    val amount: BigDecimal,
+    val amount: BigInteger,
     val description: String = ""
 )
 
 data class WithdrawMoneyCommand(
     val accountId: UUID,
-    val amount: BigDecimal,
+    val amount: BigInteger,
     val description: String = ""
 )
